+++ src/lib/gui/components/controllers/ColourBoxes.ts
@@ -1,8 +1,8 @@
 import p5 from 'p5';
 import type GUIForP5 from '../../gui';
-import ValuedController from '../../valued_controller';
+import ValuedController from '../valued_controller';
 import type { setupCallback, valueCallback } from '../../../types';
-import type Controller from '../../controller';
+import type Controller from '../controller';
 import type { P5SelectElement } from '../../../types/controller';
 
 /**
