+++ src/lib/gui/components/controllers/Slider.ts
@@ -1,6 +1,6 @@
 import type { setupCallback, valueCallback } from '../../../types';
 import type GUIForP5 from '../../gui';
-import ValuedController from '../../valued_controller';
+import ValuedController from '../valued_controller';
 /**
  * One dimensional slider controller.
  * @extends ValuedController
