+++ src/lib/gui/components/controllers/XYSlider.ts
@@ -1,7 +1,7 @@
 import type p5 from 'p5';
 import type { setupCallback, valueCallback } from '../../../types';
 import type GUIForP5 from '../../gui';
-import ValuedController from '../../valued_controller';
+import ValuedController from '../valued_controller';
 
 /**
  * Two dimensional slider returning an {x,y} object.
