+++ src/lib/gui/components/field.ts
@@ -1,5 +1,5 @@
 import type p5 from 'p5';
-import type GUIForP5 from './gui';
+import type GUIForP5 from '../gui';
 /**
  * Base GUI element container used by controllers.
  */
