+++ src/lib/gui/components/fields/Divider.ts
@@ -1,4 +1,4 @@
-import Field from '../../field';
+import Field from '../field';
 import type GUIForP5 from '../../gui';
 
 /**
