+++ src/lib/gui/components/index.ts
@@ -22,13 +22,17 @@ import Divider from './fields/Divider';
 
 import DieIcon from './DieIcon';
 
-import Field from '../field';
+import Field from './field';
+import Controller from './controller';
+import ValuedController from './valued_controller';
 
 import { Panel } from './panel';
 
 export {
 	Panel,
 	Field,
+	Controller,
+	ValuedController,
 	Title,
 	TextField,
 	ImageField,
